Moved homework from 4th lecture to a separate folder and created one for homeworks from 5th lecture
