+++ homeworks/lecture4/seventhExerciseLecture04.java
@@ -1,4 +1,4 @@
-package homeworks;
+package homeworks.lecture4;
 
 public class seventhExerciseLecture04 {
     public static void main(String[] args) {
