+++ homeworks/lecture4/sixthExerciseLecture04.java
@@ -1,4 +1,4 @@
-package homeworks;
+package homeworks.lecture4;
 import java.util.Scanner;
 public class sixthExerciseLecture04 {
     public static void main(String[] args) {
